# go api: use byte slices for deck parsing - reuse deck parsing results for subsequent reqs