--- a/backend/api/deck.go
+++ b/backend/api/deck.go
@@ -1,12 +1,14 @@
 package api
 
 import (
+	"bytes"
 	"errors"
 	"fmt"
 	"regexp"
 	"strconv"
 	"strings"
 	"sync"
+	"unsafe"
 
 	"github.com/gin-gonic/gin"
 	"golang.org/x/exp/slices"
@@ -93,9 +95,159 @@ func (d *deck) Bytes() (res []byte, err error) {
 	return d.buf, err
 }
 
-// parse
+// parse use to decompress and get readable str representation of deck, will be called once per uncompressed deck
 func (d *deck) parse() error {
+	decompressed, err := decompressBytes(d.buf)
+	if err != nil {
+		return err
+	}
+
+	parts := bytes.Split(decompressed, []byte(";;;"))
+
+	// read cards first so we can reference them as we read the card indexes
+	cards := make([][]byte, 0, bytes.Count(parts[1], []byte(";;"))+1)
+	longestCard := 0
+
+	err = readDelimitedBytes(parts[1], []byte(";;"), func(val []byte) error {
+		// TODO: some validation here
+
+		cards = append(cards, val)
+		if len(val) > longestCard {
+			longestCard = len(val)
+		}
+
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+
+	cardIdxCountMap := make(map[string]int, len(cards)) // map[card]count
+	uniqueCardList := make([]string, 0, len(cards))     // just keep copies of our strs
+
+	// now that we know which cards exist, we can read the indexes
+	err = readDelimitedBytes(parts[0], []byte(","), func(val []byte) error {
+		// compiler will optimize this string conversion out, see `cloneString` in atoi.go
+		cardIdxVal, err := strconv.ParseInt(string(val), 10, 64)
+		if err != nil {
+			return err
+		}
+
+		if cardIdxVal < 0 || cardIdxVal >= int64(len(cards)) {
+			return errors.New("card index out of bounds")
+		}
+
+		card := parseCardBytes(cards[cardIdxVal])
+		fmt.Printf("%d - %s - %s\n", cardIdxVal, string(card), string(cards[cardIdxVal]))
+
+		if len(card) == 0 {
+			return nil
+		}
+
+		// no guarentee of not copying here in map lookup, so unsafe it out. Is valid as long as `decompressed` is still valid and static
+		cardStr := unsafe.String(&card[0], len(card))
+
+		ct, ok := cardIdxCountMap[cardStr]
+		if !ok {
+			uniqueCardList = append(uniqueCardList, cardStr)
+		}
+		cardIdxCountMap[cardStr] = ct + 1
+
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+
+	// may allocate some extra space in some cases, but we will shrink after we are done formatting
+	d.buf = make([]byte, 0, len(cardIdxCountMap)*longestCard)
+
+	// put ascender's bane first
+	slices.SortFunc(uniqueCardList, func(i, j string) bool {
+		if i == "Ascender's Bane" {
+			return false
+		}
+		if j == "Ascender's Bane" {
+			return true
+		}
+
+		return i < j
+	})
+
+	for _, card := range uniqueCardList {
+		// will just use underlying bytes and not do cast
+		d.buf = append(d.buf, []byte(card)...)
+
+		cardCount, ok := cardIdxCountMap[card]
+		if !ok {
+			return errors.New("card not found")
+		}
+
+		// fmt - "$card x$count\n"
+		if cardCount > 0 {
+			d.buf = append(d.buf, ' ', 'x')
+			d.buf = strconv.AppendInt(d.buf, int64(cardCount), 10)
+			d.buf = append(d.buf, '\n')
+		}
+	}
+
+	if len(d.buf) < cap(d.buf) {
+		d.buf = d.buf[:len(d.buf)]
+	}
+
 	return nil
+}
+
+func decompressBytes(s []byte) ([]byte, error) {
+	parts := bytes.Split(s, []byte("||"))
+	if len(parts) < 2 {
+		return nil, errors.New("invalid deck")
+	}
+
+	compressionDict := bytes.Split(parts[0], []byte("|"))
+	text := parts[1]
+
+	for i := len(compressionDict) - 1; i >= 0; i-- {
+		word := compressionDict[i]
+		text = compressionWildcardRegex[i].ReplaceAll(text, word)
+	}
+
+	return text, nil
+}
+
+type delimCB func(val []byte) error
+
+func readDelimitedBytes(s []byte, delim []byte, cb delimCB) (err error) {
+	if len(s) == 0 || (len(s) == 1 && s[0] == '-') {
+		return nil
+	}
+
+	currIndex := 0
+	for currIndex < len(s) {
+		nextIndex := currIndex + bytes.Index(s[currIndex:], delim)
+		if nextIndex < currIndex {
+			nextIndex = len(s)
+		}
+		err = cb(s[currIndex:nextIndex])
+		if err != nil {
+			return err
+		}
+
+		currIndex = nextIndex + len(delim)
+	}
+
+	return nil
+}
+
+// parseCardBytes is a helper function to parse the card name from a given section
+func parseCardBytes(cardSection []byte) []byte {
+	// return first item in triplet
+	sectionEnd := bytes.Index(cardSection, []byte(";"))
+	if sectionEnd == -1 {
+		return cardSection
+	}
+
+	return cardSection[:sectionEnd]
 }
 
 func decompress(s string) (string, error) {