As listed in [this pull request](https://github.com/MaT1g3R/slay-the-relics/pull/19), the optimization can be taken a bit further.

Added deck struct to manage parsing logic.
Converted string parsing functions into their byte counterparts.
Added functionality for reusing parsed deck data between reqs.

Added tests for the bytes deck parsing logic (left the old for now, but can remove have just been using for comparisons for now).
Added tests for the deck API handler (bumped up coverage by a bit)

Do not have as good of a computer for the benchmarks, but here is the comparison from my machine:

```
goos: windows
goarch: amd64
pkg: github.com/MaT1g3R/slaytherelics/api
cpu: AMD Ryzen 7 3800XT 8-Core Processor
                              │   old.out    │               new.out                │
                              │    sec/op    │    sec/op     vs base                │
DecompressDeck/Simple_deck-16   1.947µ ± ∞ ¹   2.224µ ± ∞ ¹       ~ (p=1.000 n=1) ²
DecompressDeck/Big_deck-16      74.90µ ± ∞ ¹   59.58µ ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                         12.08µ         11.51µ        -4.68%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                              │    old.out    │                new.out                 │
                              │     B/op      │     B/op       vs base                 │
DecompressDeck/Simple_deck-16     841.0 ± ∞ ¹     554.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
DecompressDeck/Big_deck-16      220.5Ki ± ∞ ¹   116.4Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                         13.46Ki         7.935Ki        -41.04%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                              │   old.out   │               new.out                │
                              │  allocs/op  │  allocs/op   vs base                 │
DecompressDeck/Simple_deck-16   17.00 ± ∞ ¹   15.00 ± ∞ ¹        ~ (p=1.000 n=1) ²
DecompressDeck/Big_deck-16      165.0 ± ∞ ¹   112.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                         52.96         40.99        -22.61%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

Unsure of why the small deck is faster as of writing this, but the main change I wanted to implement was reuse of the request results.

Seems like even in the "large deck" cases we end up not storing that much in the decompressed buffer, was <2Kb for the large case per-item.